Add module template
Also made component template extend the module one.
