+++ templates/component.js
@@ -29,7 +29,7 @@ ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 POSSIBILITY OF SUCH DAMAGE.
 </copyright> */
 
-var TemplateBase = require("../lib/template-base.js").TemplateBase;
+var ModuleTemplate = require("./module").Template;
 var fs = require('fs');
 
 var Command = require("commander").Command;
@@ -41,74 +41,10 @@ var _fromCamelToDashes = function (name){
         return s1.replace(_allCapRe, "$1-$2").toLowerCase();
     };
 
-exports.Template = Object.create(TemplateBase, {
+exports.Template = Object.create(ModuleTemplate, {
 
     commandDescription: {
         value: "component"
-    },
-
-    addOptions: {
-        value:function (command) {
-            command = TemplateBase.addOptions.call(this, command);
-            command.option('-n, --name <name>', 'module name');
-            command.option('-e, --exported-name [name]', 'exported name');
-            command.option('-j, --jsdoc [module]', 'jsdoc module');
-            command.option('-c, --copyright [path]', 'copyright file');
-            return command;
-        }
-    },
-
-    didSetOptions: {
-        value:function (options) {
-            if (options.name) {
-                options.name = this.validateName(options.name);
-            }
-            if (!options.exportedName) {
-                options.exportedName = this.validateExport(options.name);
-            }
-            if (options.jsdoc) {
-                options.jsdoc = this.validateJsdoc(options.jsdoc);
-            }
-            if (options.copyright) {
-                options.copyright = this.validateCopyright(options.copyright);
-            }
-
-        }
-    },
-
-    validateName: {
-        value: function(name) {
-           var exportedName = this.validateExport(name);
-            // convert back from camelcase to dashes
-            return _fromCamelToDashes(exportedName);
-        }
-    },
-
-    validateExport: {
-        value: function(name) {
-            // We accept the name in any format, dashed, spaced or camelcased
-            // We then convert to to camelcase and back to get the consistent
-            // naming used in Montage
-            // remove spaces
-            name = name.replace(" ", "-");
-            // convert to camelcase
-            return name.replace(/(?:^|-)([^\-])/g, function(_, g1) { return g1.toUpperCase(); });
-        }
-    },
-
-    validateJsdoc: {
-        value: function(jsdocModule) {
-            if (jsdocModule.length && jsdocModule.substring(jsdocModule.length - 1) !== "/") {
-                return jsdocModule += "/";
-            }
-            return jsdocModule;
-        }
-    },
-
-    validateCopyright: {
-        value: function(path) {
-            return fs.readFileSync(path, "utf-8");
-        }
     },
 
     destination: {
@@ -118,7 +54,7 @@ exports.Template = Object.create(TemplateBase, {
     finish: {
         value: function() {
             var self = this;
-            return TemplateBase.finish.apply(this, arguments).then(function(result) {
+            return ModuleTemplate.finish.apply(self, arguments).then(function(result) {
                 console.log(self.options.name + ".reel created.");
                 return result;
             });
